+++ app/src/main/scala/com/waz/zclient/calling/CallingGridAdapter.scala
@@ -1,0 +1,31 @@
+/**
+ * Wire
+ * Copyright (C) 2021 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.waz.zclient.calling
+
+import androidx.fragment.app.Fragment
+import androidx.viewpager2.adapter.FragmentStateAdapter
+
+class CallingGridAdapter(fragment: Fragment) extends FragmentStateAdapter(fragment) {
+
+  override def getItemCount(): Int = 3
+
+  override def createFragment(position: Int): Fragment = new CallingGridFragment()
+}
+
+
+
